painradar/backend/requirements.txt
fastapi
uvicorn[standard]
groq
httpx
python-dotenv
pydantic
requests
